+++ packages/dash-core-components/dash_core_components/Graph.py
@@ -8,7 +8,9 @@ class Graph(Component):
 
 
 Keyword arguments:
-- id (string; required)
+- id (string; optional): The ID of this component, used to identify dash components
+in callbacks. The ID needs to be unique across all of the
+components in an app.
 - clickData (dict; optional): Data from latest click event
 - hoverData (dict; optional): Data from latest hover event
 - clear_on_unhover (boolean; optional): If True, `clear_on_unhover` will clear the `hoverData` property
@@ -84,7 +86,7 @@ so that plotly.js won't attempt to authenticate to the public Mapbox server.
 
 Available events: 'click', 'hover', 'selected', 'relayout', 'unhover'"""
     @_explicitize_args
-    def __init__(self, id=Component.REQUIRED, clickData=Component.UNDEFINED, hoverData=Component.UNDEFINED, clear_on_unhover=Component.UNDEFINED, selectedData=Component.UNDEFINED, relayoutData=Component.UNDEFINED, figure=Component.UNDEFINED, style=Component.UNDEFINED, className=Component.UNDEFINED, animate=Component.UNDEFINED, animation_options=Component.UNDEFINED, config=Component.UNDEFINED, **kwargs):
+    def __init__(self, id=Component.UNDEFINED, clickData=Component.UNDEFINED, hoverData=Component.UNDEFINED, clear_on_unhover=Component.UNDEFINED, selectedData=Component.UNDEFINED, relayoutData=Component.UNDEFINED, figure=Component.UNDEFINED, style=Component.UNDEFINED, className=Component.UNDEFINED, animate=Component.UNDEFINED, animation_options=Component.UNDEFINED, config=Component.UNDEFINED, **kwargs):
         self._prop_names = ['id', 'clickData', 'hoverData', 'clear_on_unhover', 'selectedData', 'relayoutData', 'figure', 'style', 'className', 'animate', 'animation_options', 'config']
         self._type = 'Graph'
         self._namespace = 'dash_core_components'
@@ -98,7 +100,7 @@ Available events: 'click', 'hover', 'selected', 'relayout', 'unhover'"""
         _locals.update(kwargs)  # For wildcard attrs
         args = {k: _locals[k] for k in _explicit_args if k != 'children'}
 
-        for k in ['id']:
+        for k in []:
             if k not in args:
                 raise TypeError(
                     'Required argument `' + k + '` was not specified.')